Update man page and deprecation warning for binstubs --all
Resolves #3251

Update the deprecation warning to point the user to the `--all` option
for bundle binstubs as the preferred way to install binstubs for all
gems in a bundle.

Update `bundle binstubs` man page to include the `--all` option :tada:
and remove the note pointing to the now deprecated method of installing
all binstubs.