Revert "Reactivate usused code for ruby version resolution"
This reverts commit 68929fc0af7ce40aaf2c947adb4a5d66cc229c56.

This change caused a regression if applications are using the `ruby`
Gemfile DSL, have a locked ruby version different from the running ruby,
and run `bundle update`.

In that case, the new result is something like:

```
Could not find gem 'Ruby (>= 2.2.2, = 2.7.1.83, = 2.6.5.114)' in the local ruby installation.
The source contains 'Ruby' at: 2.7.1.83
```

whereas before things would just work.

The reason is that the reverted commit started adding both the locked
ruby and the running running as exact ruby requirements when unlocking.
And they will obviously conflict if they are different.

This PR reverts the commit and adds a regression spec to catch the
error.