Fix incorrect re-resolution when path gem excluded and not available
According to the debugging messages, if specs for a path gem are not
available at `bundler/setup` time, but the group the gem belongs to has
been excluded, it should be assume that the path source has no changes
and continue.

However, there's another check for re-resolution that checks
completeness for the current platform that was not filtering for
requested groups only, and making this case to incorrectly re-resolve,
and as a result fail due to the path gem not being available (even if
not requested).

This commit should fix the problem.