Make `--default` and `--install-dir` play nice together
With this combination of flags, `rubygems` should install the default
specification to `--install-dir`, not to the default rubygems
installation.

I doubt this has any realworld use cases, but fixing it is handy for me
in order to be able to write a test for a bug fix in `bundler` related
to `bundle clean --force` removing executables of default gems.