Improve space handing when removing entries from RUBYOPT
Using `sub` to replace the entry with an empty string "" will only work
cleanly if the entry we are removing is the ONLY entry in RUBYOPT. If
it's not the only entry, it will leave behind stray whitespace.

For example, if RUBYOPT contained-r/path/to/bundler/setup
-r/something/else removing -r/path/to/bundler/setup using `sub` would
leave a leading space in RUBYOPT.

If we tried to include a trailing space when using `sub`, it would solve
that issue for when there are more than one entry but it would not work
when the entry we are removing is the ONLY entry or when it's the last
entry.

The new method of splitting on " " helps us avoid all of that because it
will "just work" for any of the above cases and it won't leave the
leading space behind in RUBYOPT. It also is consistent with the method
used to clean RUBYLIB in `unbundled_env` directly beneath this code.

The default for String#split is " " so it not technically necessary to
specify it, but it's explicitly stated here to give clarity to future
readers of this line.