Improve spec to simulate presence of -rbundler/setup in original env
Because we are trying to simulate that the RUBYOPT contained this
bundler/setup value IN the original env, we actually have to set it in
BUNDLER_ORIG_RUBYOPT for the test. We are simulating that this
bundler/setup entry in RUBYOPT existed BEFORE bundler was activated
(even though it didn't).

If we set it in RUBYOPT it will already be removed even without the code
in unbundled_env (the spec still passed when the code in unbundled_env
for handling RUBYOPT was commented out).

That is because unbundled_env calls original_env as it's starting point,
which will remove whatever we set here in RUBYOPT because it was not in
the original env when the test suite first activated bundler.