Normalize access to `DEFAULT_CONFIG`
Keys should be defined in "ENV variable format", since that's what other
helpers expect, and makes it consistent with how the other
configurations are accessed.

This fixes the issue of the `Bundler.settings.locations` hash never
returning a `:default` key. It doesn't seem to cause any issues in
practice, but still.