Test against the latest rubocop in CI
Our scripts were currently installing the exact 0.85.1 version of
rubocop, but letting rubocop dependencies be resolved to the latest
compatible version.

Sadly, `rubocop-ast 0.4.0` was released, and started being installed
here. And that broke our CI because the new version doesn't play nice
with `rubocop-0.85.1`.

Upgrade to the latest version, which should be compatible.