Fix `SystemStackError` when overriding `Warning.warn`
I changed our monkeypatch to use `prepend` instead, and that revealed an
off-by-one error in the implementation. The previous implementation was
working because (I have no idea why), two duplicated entries were
present in the callstack for `kernel_require.rb`, automatically
"correcting" the issue.

This implementation feels better and works as expected with
`Warning.warn`.