Omit deprecated gems from `gem help commands`
This change came about through discussion with David Rodgriguez. We've
been deliberating whether to document the deprecated status of commands.
I looked for examples in the OSS community and found that there's a lot
of inconsistency around documenting deprecated items. Ruby does and
doesn't do it. Rails seems to do it. Elixir seems to do it. It's not a
settled matter.

Ultimately, the decision was to remove the deprecated commands from the
`help commands` command.