Fix ruby-core intermittent test failure
It only happens randomly. I guess depending on the order tests run, the
require is sometimes provided by previous tests and sometimes not.

It can be reproduced on a copy of ruby/ruby, with

```
$ ruby tool/sync_default_gems.rb rubygems
$ make test-all TESTS="test/rubygems/test_gem_specification.rb -j2"
```

Most likely introduced by 0af1449db90748d5f484898825147f2169a7d5dd.

Fix it by requiring `date` on top of the file using it.