Fix when having `cache_all_platforms` and `specific_platform` configured
In this case, when running `bundle install`, bundler would be unable to
find some gems when updating the cache.

The reason is that the set of specs is initially materialized for the
current platform for the standard `bundle install` behavior. But then
when bundler tries to update the cache, the sources have that set of
specs cached, and bundler is unable to find specs that are not for the
current platform.

The solution is to invalidate sources cache before materializing for all
platforms, so that remote platform specific versions are considered.