Revert resolving all Gemfile platforms automatically
This is the second time I revert this change, since I have found some
cases where this might cause pain, and I want to battle test
multiplatform fixes before attempting this.

I think ideally, when we introduce something like this, we should
consider making it opt-in by recording this intention in the `Gemfile`,
maybe with an unscoped `platforms` DSL, or by an explicit CLI flag, like `bundle
install --all-platforms`, or `bundle lock --all-platforms`.

The part I'm not reverting is removing the warning about dependencies
being unused under the current platform, since it is pretty much
expected on a multiplatform context.