Fix test failure when version is exactly 3.2
In the stable branch (when version is exactly 3.2), this test fails
because rubygems updater aborts early in the specific case when we are
upgrading to the exact same rubygems version being run with an "already
installed" message.

To fix this problem, I change the test to use the earliest possible 3.2
prerelease, which is the same version that shipped the new layout and
it's a past release so it should pass consistently no matter when the
running rubygems release is.