Eureka
With the new platform specific behavior where resolving happens against
the specific platform bundler being run from, we no longer need to keep
resolving for the generic ruby platform. The specific_platform logic
already includes fallback spec groups so that we always default to the
generic platform when no more specific variants are found.

Resolving both for the generic ruby platform and the specific platform
was creating a big overhead in performance to the point where some
couldn't be resolved in reasonable time.

The ruby platform is no longer added to the lockfile unless we're using
the `force_ruby_platform` setting. To play nice with old Gemfiles
including the RUBY platform in the list of platforms while running in
frozen mode, I added some compatibility code to deal with that.

I also added a realworld spec to prove the performance improvements. With
these changes, it takes ~10s (around the same as in 2.1.4), without
these changes it hangs.