Tweak spec to be more resilient
On 2.6.6, the default version of bundler is 1.17.2. If we use 1.10.1
here, rubygems will actually activate the highest version of bundler 1.x
available, thus, the default. At least, rubygems versions before we
started doing "exact version switching".

We never want our specs to activate the default version of bundler, but
our own code. So I change the spec to use a dummy version of bundler
higher than 1.17.2.