Revert "Fix `bundle install` unintentionally saving configuration"
This reverts commit 0598cbb68c08ac47edf0d0fa6a5ec35677d3e4e5, but keeps
the same spec, but "inverted" to clarify that doing this was a feature,
not a bug.

This behavior is not present on bundler 3 though, because it won't
remember CLI options, and what the current code does is to fake `--with`
and `--without` where given as CLI options if they are configured
globally.  We don't want to preserve this behaviour on the next major
version, so the same spec that was present on master (setting are not
automagically saved from global to local config) is kept on bundler 3.

In addition to that, I'm adding another spec representing the realworld
behavior that was broken for people when I introduced
0598cbb68c08ac47edf0d0fa6a5ec35677d3e4e5.

Note that this spec does not currently pass on bundler 3 (again, because
bundler 3 does not remember CLI options automatically), but I plan to
fix that by changing `bundle config` default to save configuration
locally to the application, not globally.