Fix incorrect path to rubygems being passed to rake builder
We were passing an incorrect folder, and one which has a `openssl.rb`
file inside. So this file was shadowing the require of the "real openssl".

This `-I` parameter is not really needed in real life I believe, since there can
be at most one rubygems installation at a given ruby installation. I think the
idea of this parameter is that our tests never leak to the "system rubygems",
because when testing there are two copies of rubygems involved (the one being
tested, and the one from the ruby we're running).

I'm not sure using this `-I` parameter here is a good idea, so since I'm not
sure I'm just keeping things as they are for now and fixing the bug.