Move exception handling around the place that should raise
We currently have a CI failure that I'm not sure where it's being
originated. By moving the exception handling to a more specific place
we'll get either a full error with a backtrace if the exception is not
happening around this place, or the same thing if it's the subcommand
generating it somehow.