`DepProxy.get_proxy` factory function
This means that all equal `DepProxy` objects will be the same object as
they are deduped on create. This then in turn allows
`Bundler::Molinillo::Resolver#build_details_for_unwind` function to use
`Array#-` to find relevant unused unwinds which is about 5 times faster
than `Array#&` was before. This dramatically improves performance for
some use cases as this function was where most of the time in the resolution
was being consumed.