Fix running `rake release` from an ambiguous ref
If the user has a tag and a branch with the same name, `git rev-parse
--abrev-ref HEAD` returns a qualified ref like `heahs/v1.1.0`. And then
`git config --get branch.<name>.remote` receives an incorrect branch
name and things fail.

There's a newer `git branch --show-current` that displays the current
branch name without qualified references, but it's git v2.22.0+ only, so
I just left a note in there in case we want to use that in the future.

Finally, I also had to migrate the reference to the explicit tag being
pushed because because of the same reason: it could be ambiguous.