Introduce disable_local_revision_check config
This config, in combination with disable_local_branch_check make it
so `local.library` configuration work in the same way as changing
them Gemfile to point to a path.

This is useful specially in bisect scripts, which will change the HEAD
of the repository to point to a commit that will not include the commit
specified in the Gemfile.lock.