Fix `gem outdated` incorrectly handling platform specific gems
If you have an up to date platform specific gem installed, and a lower
ruby specific version, it would report the gem as out of date.

This commit changes the outcome of `Gem.latest_specs` to return the
latest spec for each gem, not the latest spec for each gem and platform.

I don't see how the previous behavior could be useful, and it was
causing this bug.