Show backtrace for all potential flakies
I added the `DEBUG` env to one spec that failed some days ago. A new
failure happened in another spec. The new failure confirms my suspicion
that the culprit is in the rake extension compilation.

I have a potential fix, but I want to confirm the culprit first, so
I'm showing a backtrace by default in all specs that use the offending
dummy spec.