Fix spec to not use a custom path
If a `--path=<custom_path>` is given to bundle cache (or if BUNDLE_PATH
is configured), `bundle cache` will install the cache to
`<custom_path>/vendor/cache`. This means that in this case the cache is
being installed to `./vendor/cache/vendor/cache`, which is unexpected.

The only reason this spec is working is that currently `bundle cache`
installs the cache twice, which is a bug that I'll fix later.