Enable `disable_multisource` functionality by default
We have several source priority bug fixes in master that we have never
released because they need the `disable_multisource` setting toggled, and
that will only happen by default in Bundler 3.

The reason for this is that this setting drops support for multiple
global sources in the `gems.rb` file, and also enables separate sections
for each rubygems source in the lockfile. We considered this to be
backwards incompatible, although I'm not sure that was the right call
after looking into this.

So this commit enables the functionality by default, and completely
removes the `disable_multisource` setting. The only backwards compatible
concern I'm having is when some is using a lokfile in the old format in
frozen mode. In that case, to not break any workflows, we print a
warning and still use that lockfile. Also, plugins still need the
previous functionality. I think this can probably be fixed later, but
I'm not doing that here.

In every other case, we will run in the new secure mode and update the
`Gemfile.lock` file with the new format.