Add a regression spec to cover manageiq case
In 2.2.10 we introduced an issue where we would incorrectly print a "gem
available in multiple sources" error.

Cover that case to make sure we don't regress there.

Note that the error is still introduced if you enable the
`disable_multisource` setting, but we haven't yet enabled that behavior
by default, so we're fine for the moment.