Enable separate lockfile GEM sections by default
Previously bundler would merge all rubygems sources into a single GEM
section in the lockfile. That means that it's non deterministic which of
those source each gem is pulled from, and new releases in those sources
could change it. Now once you bundle once successfully, you're sure you
have a deterministic gem => source mapping, and no releases on any
source can change that, unless you explicitly change/unlock your
sources.

In 2.2.10, this change introduced an issue where `bundle install` would
unlock all sources and their gems by default, and thus would endup
re-resolving and installing the latest version of gems, not the locked
one.

This is somewhat inherent to the change of locking GEM sources
separately in the lockfile. If we change the lockfile format, the list
of sources changes, so we can't lock the previous specifications since
they come from an expired source.

I guess we could introduce a new mechanism to say "unlock the source of
this specification, but not its version". But it sounds like a
non-trivial feature with a very limited usage other than this migration.
So, the approach I took here is to keep the old format for existing
lockfiles and print a warning that it's not safe, so that people still
get `bundle install` working as before, but get a warning that they
should run `bundle update` or completely remove the lockfile and
regenerate it from scratch.