Run hooks tests on gemspecs not already installed
The current `setup_base_installer` ends up using the `quick_gem` helper,
which leaves the created specification installed. Instead, make sure to
use the `util_spec` helper, which does a similar thing but doesn't leave
the specification installed.

The idea is that tests do not rely on the installer removing existing
gemspecs, bacause I plan to stop doing that.