Fix deployment install with duplicate path gems added to Gemfile
When adding the same path gem twice to a Gemfile, bundle install will 
succeed with deployment set to "false" but fail with deployment set to 
"true". The error message given is "The gemspecs for path gems changed". 
To fix, duplicate dependencies must be removed before checking against 
the locked dependencies.