Add standard option alongside rubocop to `bundle gem`
This commit adds a new --linter option. The option supports either the standard or rubocop gems,
and (similar to the --ci option), it can also be set either interactively or via a new
`gem.linter` global config.

This leaves in place the --rubocop and --no-rubocop options and `gem.rubocop` config property,
but only uses them in the absence of a linter setting (and when present, prints a deprecation
warning.

The gemspec template is also updated so that empty projects will pass both rubocop and
standard's default styles (so that rake will exit cleanly for users regardless of linter setting)