Pass target folder directly to `git init`
I want to move repo initialization earlier on, so that I can figure out
the default branch in the generated repository _before_ rendering the
README template, so that I can use that information there.

At that point, the target folder might not yet exist, so by passing it
as an argument to `git init` we make sure it's created automatically if
it doesn't yet exist.