Don't try to install non installable gems if we can avoid it
If the lockfile is only locked to RUBY, we take that as a sign that it
was either generated with an old version that didn't lock platforms, or
that the user is intentionally not locking platforms. That means we need
to run in "legacy mode" and do platform resolution "lazily" at
materialization time.

This late resolution does not consider metadata requirements, so it can
happen that bundler chooses a platform specific variant not compatible
with the running ruby, when the pure ruby platform is.

To fix the issue, we need to consider metadata requirements in this late
resolution step.

Unfortunately, if running in the "new mode" with a lockfile including
specific platforms, this late resolution step doesn't happen and only
information from the lockfile is used, and there's no lazy platform
resolution at all. So to fix this issue in that case we'll need to
include metadata requirements in the lockfile. But that's for another
time.