Cancel `deployment_means_frozen` change
We have the following change scheduled for bundler 3:

* Remove the `frozen` setting.
* Change the current `deployment` setting to mean what the `frozen`
  setting means today.

I think this is not a good change, because:

* It does not have a clean deprecation/migration path: we can't
  deprecate frozen and give an alternative, because `deployment` does
  not an work today as it will work in bundler 3.

* When removing `frozen` was proposed, one of the motivations was that
  there was no `deployment` setting, only a `--deployment` flag. Today
  we do have a `deployment` setting so that makes the migration path
  harder, and the change kind of unnecessary.

I do understand that a `deployment` setting that means `frozen` +
`path=vendor/bundle` might be kind of obscure/unnecessary. But our users
find it useful and it's used all around the ecosystem.

I think we have sufficient breaking changes for bundler 3, like changing
bundler to install locally instead of globally, so I prefer to cancel
this one.