Make sure new dependency names are properly fetched
When running `bundle cache` with excluded groups and a pre-existing
lockfile, first `bundle install` is run excluding gems in the excluded
groups. After that, the whole set of gems should be cached including
those from excluded groups.

In the case where the lockfile is up to date and specs are not
previously installed, we're only dealing with `LazySpecification`
objects built from information in the lockfile. These need to be
materialized into real specifications that can be fetched.

Materializing these specs happens twice in this case, first when
fetching gems for installation, then when fetching gems for caching. In
the case of installation, we don't want to fetch excluded gems, but when
caching we want to fetch them. This means the second time we materialize
the set of specs, the dependency names to be fetched from the main
source have changed, and we need to make sure to fetch any new
specifications that might be missing. It sounds like this is what
setting `dependency_names` was trying to do here, but this has no effect
on spec fetching. There's a `double_check_for(names)` method to do just
that.