Do not use openssl constants in Gem::Specification#to_ruby
The attribute can be either a String indicating the path of the private
key file or a raw key object (an instance of OpenSSL::PKey::RSA).
Currently, the attribute is excluded from #to_ruby in the latter case.

The current code does not interact well when the openssl library is
missing. Evaluating "defined?(OpenSSL::PKey::RSA)" will attempt to
activate the openssl gem due to autoload, and that causes weird test
failures.

Since we know what type signing_key will be otherwise, which is String,
we can simply check that instead.