Double check for missing names last thing
Currently, `source.remote!` and `source.local!` reset the memoized specs
of the source, which means that remote specs are picked again from the
originally fetched specs. That means that if double checking made any
additions over that, they are cancelled unless this "double checking" is
that last thing we do.