Better explain how to run tests
It's a shame that it's so hard to find out how to run tests without running the entire test suite.

In a project like this, when making a contribution, it's often unnecessary to run the entire test suite when the
CI will run it regardless on every commit, it's much more helpful to new contributors to explain how individual test
files they may be affecting can be run, and even individual test methods, since it's not discoverable through any of the
feedback from Rake when running `rake test`.