Fix isuue with cached gems with extensions being ignored
Cached gems are not really installed, so checking whether the gemspec
they package is missing extensions makes no sense. This was accidentally
working before because old rubygems versions included a check like the
following

```ruby
def missing_extensions?
  # ...
  return false if installed_by_version < Gem::Version.new("2.2.0.preview.2")
  # ...
```

In the case of these cached gemspecs, the `installed_by_version`
attribute is never set since the gemspec is not really installed, it's
just extracted from the cached `.gem` package. So it gets the default
`Gem::Version.new(0)` value, and the `missing_extensions?` check was
always false.

This condition was however removed in rubygems 3.2.18, making these
specs incorrectly skipped.

The solution is to remove the check, since it doesn't make sense for
cached gems.