Stop testing ruby-core CI in the stable branch
Ruby-core CI is to test the integration of the latest HEAD of ruby-core
with the latest HEAD of rubygems. It doesn't make sense to run this on
the stable branch. If could make some sense to run the integration
against a manual build of ruby's stable branch, but not against a manual
build of ruby-head. Since there are no such builds available, I think
it's best to skip this job in the stable branch.

The main problem with running this workflow on the stable branch is that
it's hard to get green and sometimes requires backporting development
changes from master, which increases the risk of conflicts which need to
be manually fixed by backporting other dependent PRs or by writing
custom code. Not running it here should make our releases smaller, yet
passing the stable branch's CI.