`bundle exec --allow-gems=GEMS` to enable GEMS regardless of Gemfile
This changeset adds `--allow-gems=GEMS` option to `bundle exec`.
The option makes the speficied gems available even if the gems are not
listed in Gemfile. This is supposed to be used for some tooling gems
(like stackprof) which is needed just temporarily.

If the gem specified by `--allow-gems` is not listed in Gemfile, the
installed latest version of the gem is picked up. If the gem is listed
in Gemfile, the option is just ignored and Gemfile's specification is
respected.

Fixes #4649