Fix multisource lockfile warning message
Recently we fixed our logic to automatically update the lockfile format
to have split GEM sections when necessary. So, there's only two cases
when this warning is printed now:

* When the Gemfile has multiple global sources. In this case, the
  warning is not really necessary because we already display a deprecation
  warning about the `Gemfile`, and once that's fixed, the lockfile format
  will be automatically upgraded.

* When we are in frozen mode, because we're not allowed to write to the
  lockfile.

This commit changes bundler so that the warning is not printed in the
first case, and so that the warning gives proper resolution steps for
the second case.