Allow setting --otp via GEM_HOST_OTP_CODE
For release automation it can be nice to want to pass in the OTP via
some UI (e.g. a workflow_dispatch input on GitHub Actions) and then just
be able to run `rake release` to publish to rubygems.

This introduces a new environment variable `GEM_HOST_OTP_CODE` which
will be checked if the `--otp` command line option is not set, allowing
for automated releases with OTP.