Don't require bundler absolutely when running specs
All specs except the ones using this matcher require bundler without an
absolute path like it's done in the realworld. By doing that, we also
make sure rubygems code that activates the proper bundler version is
also exercised.

This commit changes this matcher to be consistent with the rest of the
specs.