Ensure uniqueness when adding rubygems sources
If there's duplicated global and scoped sources, it can happen that we
end up adding two different sources that are actually the same, due to
the way and order sources are added.

To avoid this, we make sure each source is added immediately when it's
defined, and that we check for duplications before adding them.