Delay cache access in `LockfileParser`
It's the only part that needs "root folder resultion" to figure out the
folder for the cache, but it's only needed for some things, so run that
logic lazily when needed.

The motivation for this refactoring is a bit complicated:

* In our development environment, we use `LockfileParser` to figure out
  the exact version of each development dependency we need to load. We
  could do this via `bundler/setup` but we have never got that to work for
  bundler specs themselves (use bundler to test bundler).

* Our usage of `LockfileParser` relies on the presence of a
  `.bundle/config` file in the root of the repo, otherwise "root folder
  resolution" would fail to find both a `Gemfile` or a `.bundle/config`
  file and would fail.

* An upcoming bug fix to make `bundler/setup` also consider gemfile
  configured through `.bundle/config`, will make it so that the
  `.bundle/config` file at the root of this repository will be used by
  bundler development tasks, and that will break things.

So I want to remove the `.bundle/config` file, but to do that I need to
make `LockfileParser` not do root folder resolution by default.

Hence, this change.