Adapt spec to explicitly pass the `--global` flag to `bundle config`
If we make `bundle config` local by default, this spec will no longer
work because it's relying on a `bundle config set path` command to apply
also to a folder completely outside of the folder hierarchy where the
`bundle config` command is run.

The spec is verying a completely related feature, and was just working
by chance due to the fact that `bundle config` is currently global by
default.