Ensure vendoring gems downloads licenses
Bundler vendors gems using Automatiek (itself a forked, vendored
version). Most of these vendored gems have licenses which require
the license itself to be distributed with the sourcecode.

Until this commit, licenses were not downloaded and added to this
repository as part of vendoring gems.

It is worth noting that this change causes the vendored copy of
Automatiek to diverge from the upstream. This is already the policy of
bundler.

Co-authored-by: Frederik Dudzik <frederik.dudzik@shopify.com>
Co-authored-by: Jacques Chester <jacques.chester@shopify.com>