Remove unnecessary mutexes
These were added initially to allow for concurrent `gem install`
commands, but these are not really supported anyways, because the
installer first completely wipes out existing installations, and then
unpackages files, which is obviously not thread safe. We would need a
file lock or something to make it thread safe, but no one has ever
complained about it so I don't think it's necessary. In any case, I
don't think the current mutex is effective so I'll remove it.