Load `operating_system.rb` customizations before setting up default gems
It's very common for packagers to configure gem paths in this file, for
example, `Gem.default_dir`. Also, setting up default gems requires these
paths to be set, so that we know which default gems need to be setup.

If we setup default gems before loading `operatin_system.rb`
customizations, the wrong default gems will be setup.

Unfortunately, default gems loaded by `operating_system.rb` can't be
upgraded if we do this, but it seems much of a smaller issue. I wasn't
even fully sure it was the right thing to do when I added that, and it
was not the culprit of the end user issue that led to making that
change.

So as part of this change we also need to:

* Revert the changes that led to adding the above functionality.
* Update some the version of fiddle installed by specs that need to
  customize `Gem.default_dir`.

For completeness, the explanation of why we need to do the latter, it's
because these specs were installing a version of fiddle to this
customized default dir that no longer matches the default fiddle version
on ruby 3.0.3. That means that while before this change, fiddle would
required after the default dir had been changed (and whichever version
installed there would be correctly picked up), now fiddle is required
and activated as a gem _before_ this folder is customized, and then will
fail to be found when eagerly resolving dependencies from bundler's
binstub. So we now need to install a version exactly matching the
highest default version in our support matrix to avoid that. This is
still brittle, but I won't be further addressing that here.