Fix some test commands losing `ENV["RUBYOPT"]`
If an explicit RUBYOPT was passed to our test command helpers, the
global RUBYOPT in the environment would be lost. This seems important on
Windows since it has a global `-Eutf-8` that seems to cause issues when
lost.

This commit fixes that by properly merging both the explicit environment
given and the one globally set.