Prepend `Gem.ruby` if `$PROGRAM_NAME` is not executable
Most of the times, `bundle` should be an executable file with a shebang
that gets run directly. However, in some edge case, like our development
environment for bundler itself, this is not true.

This commit prepends `Gem.ruby` to the rexec'ed command line when we're
not in the standard case.