Use bundler installer to install the bundler gem
This is how bundler installs gem, so we should use the same way for the
bundler gem itself. In particular, Bundler does not use `Gem::Installer`
but a subclass of it that does not call `Gem::Specification.reset` after
install, and that also does some checksum verifications.

This also allows us to:

* Properly verify that the `BUNDLED WITH` version actually exist.
* Not depend on what's on `Gem.sources`.