Fix metadata requirements being passed for custom gem servers
Custom gem servers usually implement only the old dependency API, which
provided no metadata requirement information. That means those
requirements would be unconstrained in practice for these servers, and
result in potentially installing gems not compatible with the running
Ruby or RubyGems version.

To fix the issue, fallback to picking up this information the marshalled
index when the old dependency API is used, and thus we have no metadata
information available.