Don't consider gems provided by Debian packages as deleted
In Debian, old-style packages provide a gemspec but the actual library files
are shipped in /usr/lib/ruby/vendor_ruby/. We are converting all packages to
the Rubygems layout, but we are not finished yet.

This fixed a regression introduced in Debian by
8950631f02498195794096a8e7c28b7feac56382 ("Fix `bundle install` to force
reinstallation of deleted gems").

Bug-Debian: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1004195