Don't mutate spec sets on access
Before, every time we accessed a spec set with a name not present in the
underlying hash, an entry with an empty set of specs was created for it.
I think it's best to assume all keys in a spec set have actual keys, at
least I was assuming that until I found about this.