Fix test errors when `cargo` not present
Currently our tests try to detect whether `cargo` is installed or not,
and if not, set tests that need `cargo` as pending.

However, when this happens that test `setup` method is completely
skipped, meaning that the `teardown` method will blow up when trying to
switch back to the original folder, since it was not set.

This commit fixes that.