Improve Ruby version resolution errors
They are not very accurate because the always mention the "Ruby" version
is set in the `Gemfile`, but that's not always the case.

Also, when `ruby` not set in the `Gemfile`, they fail to mention the
current Ruby version, which is what's causing the actual conflicts.

The `ruby` Gemfile directive is already validated against the current
Ruby, so we can pass the current Ruby directly to the resolver. And
mention "Current Ruby version" instead of "In Gemfile" in resolution
errors.