Remove unnecessary `Lint/UselessElseWithoutRescue` cop
The offenses detected by this cop are actually a syntax error on newer
rubies, so it's very unlikely that they would go unnoticed so we
extensively test newer rubies. Due to this, newer versions of RuboCop
have removed it, so I'm removing it too to play nice with those
versions.