Fix `bundle package --no-install` no longer skipping install
This is a regression from cf749f8ffabdfb60cf9c8e17689199e6ea45b6e2. The
funny thing is that we have a spec for this feature, so it was unclear
how we regressed here. It turns out there was a bug in one of our
negative matchers checking that gems ARE NOT included in a bundle.

This commit fixes the bug in the negative matcher and reverts
cf749f8ffabdfb60cf9c8e17689199e6ea45b6e2 (with a slightly simpler diff).