Document how to run rake setup under regular user
When running rake setup under regular user, it fails with permission
errors since it tries to install into global gem directory. Running it
under root is a workaround, however in that case bundler complains that
it should not run under root.

Setting GEM_HOME is a working solution, so it should be documented.