Rename `SpecSet#spec_for_dependecy` to `SpecSet#specs_for_dependency`
It's more correct, since an array of specs is return, but also
workarounds an issue in truffleruby, which will forcefully patch all
Bundler versions including this method in order to cherry-pick a bug
fix rather than patching only the versions where the bug is present.
That means that truffleruby will ignore any other bug fixes /
improvements this method receives in the future. An easy escape to this
is to rename the method.