Make sure truffleruby properly deals with old lockfiles
Until 2.3.17, Bundler would always use the RUBY platform for Truffleruby, so
the RUBY platform would always be used in lockfiles generated on Truffleruby.

From 2.3.18, Bundler will consider `Gem::Platform.local` as
Truffleruby's platform, just like CRuby, and use that inside lockfiles.

But we need to properly deal with old lockfiles that used the RUBY
platform.

This commit implements that.