Don't remove unnecessary gems from lockfile unless necessary
Doing this has a cost for every `bundler/setup` invocation, however,
it's not necessary in most cases. It's only needed if:

* Resolution includes a pretty convoluted combination of platform
  specific dependencies.
* Lockfile was generated with Bundler < 1.16.2, and was re-bundled
  with a newer version.

So, we can detect the above situation and only do the cleanup in that
case.

On a reasonably big example Gemfile, this provides a `bundler/setup`
speed up of about 5%.