Fix Bundler no longer installing as expected on TruffleRuby
If the lockfile had been generated on another platform, we recently lost
the ability to install the proper gems on TruffleRuby.

The way I can think of to keep our current cleaner approach of using
`force_ruby_platform` would be that TruffleRuby was reported as a
special platform. But that would require further discussion, so for now
I'm reverting the changes that caused this.

In particular, reverts "Make sure truffleruby properly deals with old
lockfiles" (the "lib" part of commit
f9710051efa4f25812b8c118aea74e3bdf9d3b5b) and "Better solution to
support truffleruby platform handling" (the "lib" part of commit
6085de09c711f072ddd9fa06d125a6da65ebc501).