Fix yet another TruffleRuby platform selection regression
I got this bug reported before the 2.3.18 release. I reproduced it and
confirmed that recent changes in master had already fixed the bug and
also that we had a spec testing explicitly this behaviour, but noticed
that the bug went uncaught because we were only testing on TruffleRuby
22.0 (the spec did not regress there), not on TruffleRuby 22.1, where
the spec did start failing when the regression was introduced.

So I called the bug fixed and upgraded our CI to use the latest
TruffleRuby 22.1.

However, later I noticed another related bug that required reverting
some changes aimed for TruffleRuby to fix it. Well, reverting those
changes made the bug reappear again, but our spec did not catch it and
apparently I did not test it in real life.

The reason the spec did not catch it is that due to how the spec was
written, it would not reproduce the original scenario if
`Bundler.local_platform` returns "ruby". Because in that case, the most
specific variant available of the dummy spec on Linux would be "ruby",
so we would no longer be testing the more complicated scenario where a
platform specific variant that MUST NOT be chosen. And unfortunately the
changes reverted made `Bundler.local_platform` return "ruby" again on
TruffleRuby.

This commit:

* Fixes the spec to use `Gem::Platform.local` instead of
`Bundler.local_platform`, since that always returns the "real specific
platform", so that test will always include a platform specific variant.

* Fixes the bug with the opposite change, make sure Bundler uses "ruby"
as the local platform for TruffleRuby (and thus does not pick platform
specific variants).