No need to deal with dependencies in index
Just names.

This changes also fixes an issue where some dependencies would be
removed from the lockfile when global gemspecs have incorrect
dependencies. Now the lockfile dependencies take priority.

This is due to a ruby-core issue, but I'm covering it with a spec to
avoid getting hit again by it.