Don't add duplicated spec groups
Adding duplicated spec groups for both platform specific and ruby
specific groups makes the number of candidates duplicate all gems that
end up considering both. Our criteria for choosing the next dependency
to process is based, among other things, on the number of candidates for
a particular dependency, so artificially doubling it causes issues where
the wrong dependencies are considered before others.

To implement this,  we start comparing heterogeneous spec types, and
that break a particular case of yanked gems errors, so add a spec to
cover that.