Fix bad unfreeze recommendation
If failing because lockfile changed while in frozen mode, Bundler
recommends to run a command in development to temporarily unfreeze
Bundler and regenerate the lockfile.

That recommendation does not make sense, however, if Bundler is frozen
through ENV, and it currently results in a meaningless advice reading:

```
If this is a development machine, remove the /path/to/Gemfile freeze by running ``
```

This commit fixes the issue by avoiding displaying any recommendation in
this case.

Co-authored-by: Josef Šimánek <josef.simanek@gmail.com>