Properly sort specs when materializing
When a Gemfile.lock file is locked only to the "ruby" platform, we still
do some resolution of the best platform to install when materializing
specs. But these specs come from the API without any specific ordering,
so since we recently stopped always sorting specs, we need to make sure
the proper priority is used here, so that we always prefer a precompiled
gem if available.