Do not fill the template with the gem name automatically
This PR comes as part of our security work. With @simi, @indirect, and @hsbt  an increase in brandjacking attacks where the brand's origin comes from scanning GH for names defined in the README.md but not in RubyGems.

It seems that often the default flow for people is to start their gems and upload them to GitHub without actually releasing them to RubyGems. In such cases or in cases where the author does not want to publish to RubyGems at all, the package name can be taken over and used for malicious activities, hoping someone will follow along with the recommendation from the README to install from RubyGems.