Fix crash due to `BundlerVersionFinder` not defined
This seems like some kind of autoload Ruby bug, since we do setup
autoloads for both `BundlerVersionFinder` and `Dependency` at RubyGems
boot.

The autoloading of `BundlerVersionFinder` should be happening during the
`autoload` of `Dependency` and I guess that's something Ruby cannot
handle?

I'll try to bring this to ruby-core but the easy fix for us is to remove
the `BundlerVersionFinder` autoload and move it to an inline `require`
which should have the same effect.

Co-authored-by: Joshua Swett <joshuaswett@users.noreply.github.com>